Move implementation to templates file
